binding posts to users

+++ Application/Controllers/PostController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 using MvcPost.Models;
 using MvcUser.Data;
 using MvcUser.Models;
@@ -21,7 +22,8 @@ namespace MvcPost.Controllers
         // GET: /Post/ 
         public IActionResult Index()
         {
-            return View();
+            var user = User.Identity.Name;
+            return Content(user);
         }
 
         //GET: /Post/List/
@@ -30,7 +32,7 @@ namespace MvcPost.Controllers
             return View(await _context.Post.ToListAsync());
         }
 
-        //GET: /Post/Details/
+        //GET: /Post/Details/1
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -56,13 +58,17 @@ namespace MvcPost.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(User user, Post post)
         {
-            
+            DateTime thisDay = DateTime.Now;
+            post.Date = thisDay;
+            Console.WriteLine(thisDay.ToString());
+            post.UserId = Int32.Parse(User.Identity.Name);
+
             _context.Add(post);
             await _context.SaveChangesAsync();
             return View(post);
         }
 
-        //GET: /Post/Edit/
+        //GET: /Post/Edit/1
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -78,7 +84,7 @@ namespace MvcPost.Controllers
             return View(post);
         }
 
-        //POST: /Post/Edit/
+        //POST: /Post/Edit/1
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, Post post)
@@ -89,11 +95,15 @@ namespace MvcPost.Controllers
             }
 
             if (ModelState.IsValid)
-            {
-                _context.Update(post);
-                await _context.SaveChangesAsync();
-                
-                return RedirectToAction(nameof(Index));
+            {   
+                if(post.UserId.ToString() == User.Identity.Name){
+                    _context.Update(post);
+                    await _context.SaveChangesAsync();
+                    
+                    return RedirectToAction(nameof(Index));
+                }else{
+                    return NotFound();
+                }
             }
             return View(post);
         }
@@ -119,10 +129,15 @@ namespace MvcPost.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Delete(int id)
         {
+            
             var post = await _context.Post.FindAsync(id);
-            _context.Post.Remove(post);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if(post.UserId.ToString() == User.Identity.Name){
+                _context.Post.Remove(post);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }else{
+                return NotFound();
+            }
         }
         
         //GET: /Post/MyList/
